Add files via upload
A calculator is a device that performs arithmetic operations on numbers. The simplest calculators can do only addition, subtraction, multiplication, and division.

This GUI is made by the tkinter module of python and use eval function for solving all arithmetic operations. this is made by the grid format specifier in this GUI we use a single Label for displaying the inputs and output and makes buttons for giving the command i use many function or arguments in this calculator like Justify and Relief etc.